Update signal-message-exporter.py
This is needed to fix a major bug. I got a little carried away with find and replace when updating from Signal's old database format, which had changed the name of the 'msg_box' column to 'type'. The XML still needs to have the msg_box attribute or MMS messages sent from the user's phone will not reflect the correct sender.